[AGENT-209] Refactor adding env vars from file
